Mise en place du systeme avec le file_reader_resul

+++ my_types.h
@@ -5,7 +5,7 @@
 typedef struct {
     int nb_iterations;
     int taille_vecteurs;
-    int ** vecteurs;
+    double ** vecteurs;
 } file_reader_result_t;
 
 #endif
